s2: + CLI args parser
