+++ src/sorting/InputType.kt
@@ -1,0 +1,7 @@
+package sorting
+
+enum class InputType {
+    NUMBER,
+    LINE,
+    WORD
+}
